refactor(processor): extract page extraction
Move per-page text, geometry, annotation, and visual signal assembly into a dedicated processor helper so processDocument reads as the document-level pipeline.
