+++ src/core/processor.ts
@@ -12,23 +12,16 @@ import type {
   PageResult,
   ProcessDocumentOptions,
   ProcessOptions,
-  TextSpan,
   VectorBox,
 } from '../types/index.js';
-import { buildAnnotations, hasVisibleAnnotationAppearance } from './annotations/index.js';
 import { buildAttachments } from './document/attachments.js';
-import { resolveDestinationPage } from './document/destinations.js';
 import { buildLayers } from './document/layers.js';
 import { buildOutline } from './document/outline.js';
-import { buildPageStructure, countStructureNodes } from './document/structure.js';
-import { buildViewerState, normalizeJavaScriptActions } from './document/viewer.js';
-import { buildFormFields } from './formFields/index.js';
-import { buildImageBoxes, type ImageOps } from './graphics/imageBoxes.js';
-import { buildVectorBoxes } from './graphics/vectorBoxes.js';
-import { countVectorPaintOps } from './graphics/vectorOps.js';
+import { countStructureNodes } from './document/structure.js';
+import { buildViewerState } from './document/viewer.js';
+import type { ImageOps } from './graphics/imageBoxes.js';
 import { getCacheDir, getCached, pdfFingerprint, setCache } from './io/cache.js';
-import { buildLayout, markRepeatedBlocks } from './layout/index.js';
-import { buildLinks } from './links/index.js';
+import { markRepeatedBlocks } from './layout/index.js';
 import { parsePageRangeWithSkipped } from './options/pageRange.js';
 import { buildCacheKey } from './processor/cacheKey.js';
 import {
@@ -37,335 +30,18 @@ import {
   dropCachedSafe,
   isUsableImage,
 } from './processor/cacheValidation.js';
-import type { PageData, PageFlags } from './processor/pageData.js';
+import type { PageFlags } from './processor/pageData.js';
+import { extractPageData } from './processor/pageExtraction.js';
 import { fingerprintData, withTruncationHint } from './processor/pdfBytes.js';
 import { prepareRenderImagesDir, validateRenderRegion, validateRenderScale } from './processor/renderOptions.js';
 import { renderResult } from './processor/renderResult.js';
-import { normalizeText, round2, textItemDedupeKey } from './processor/textUtils.js';
-import { nonPrintableStats } from './quality/nonPrintable.js';
+import { normalizeText } from './processor/textUtils.js';
 import { derivePageQuality } from './quality/pageQuality.js';
-import { isRasterBackedTextLayer } from './quality/rasterBackedTextLayer.js';
 import { runParallel } from './runtime/parallel.js';
 import { type CompiledSearch, compileSearch, searchPage, suppressDuplicateOcrMatches } from './search/index.js';
-import { type JoinItem, joinPageText } from './text/cjkJoin.js';
-import { textMatrixFontSize, textRunGeometryFromTransform } from './text/geometry.js';
 import { type BuildVisualRegionsInput, buildVisualRegions } from './visualRegions/index.js';
 import { detectPageWarnings } from './warnings/index.js';
 import { extractWidgetAppearanceCaptions } from './widgetAppearance/index.js';
-
-function hasPushButtonWidget(annotation: unknown): boolean {
-  if (!annotation || typeof annotation !== 'object') return false;
-  const raw = annotation as { subtype?: unknown; fieldType?: unknown; pushButton?: unknown };
-  return raw.subtype === 'Widget' && raw.fieldType === 'Btn' && raw.pushButton === true;
-}
-
-/**
- * Extract a page's text plus rough density metadata.
- *
- * `imageCount` and `textCoverage` let agents detect "looks fine but the
- * real content is rasterised" pages (common in Google Slides exports)
- * and decide whether to re-run with `--render`.
- */
-async function extractPageData(
-  doc: PDFDocumentProxy,
-  pageNum: number,
-  ops: ImageOps,
-  flags: PageFlags,
-  getWidgetAppearanceCaptions?: () => ReadonlyMap<string, string>,
-): Promise<PageData> {
-  const page = await doc.getPage(pageNum);
-  const content = await page.getTextContent({ includeMarkedContent: true });
-  const optionalContentText = content.items.some(isOptionalContentTextMarker);
-
-  const view = page.view;
-  // MediaBox is normally [minX, minY, maxX, maxY] but the spec allows the
-  // pairs in either order; use abs so a flipped box still yields a sensible
-  // area instead of falling through to 0 coverage.
-  const width = Math.abs(view[2] - view[0]);
-  const height = Math.abs(view[3] - view[1]);
-  const xMin = Math.min(view[0], view[2]);
-  const yMin = Math.min(view[1], view[3]);
-
-  // Spans are the input to layout reconstruction AND to search bbox
-  // computation, so we build them whenever any of those needs is set —
-  // even though we may only expose them on PageResult when `geometry`
-  // is on.
-  const wantSpans =
-    flags.geometry ||
-    flags.layout ||
-    flags.visualRegions ||
-    flags.formFields ||
-    flags.links ||
-    flags.needSpansForSearch ||
-    flags.needFormFieldsForSearch;
-
-  // Collect typed items for the CJK-aware page-text joiner. We can't
-  // build the final string in this loop because the join decision for
-  // a whitespace item depends on its neighbours' positions, which we
-  // only know after the walk.
-  const joinItems: JoinItem[] = [];
-  let textArea = 0;
-  const spans: TextSpan[] = [];
-  const seenTextItems = new Map<string, number>();
-  for (const item of content.items) {
-    if (!('str' in item)) continue;
-    const w = typeof item.width === 'number' ? item.width : 0;
-    // pdfjs reports item.height as 0 for many PDFs (e.g. those produced by
-    // certain Office exporters); fall back to the vertical scale from the
-    // text matrix, which is effectively the glyph height in user units.
-    const reportedH = typeof item.height === 'number' ? item.height : 0;
-    const transform = item.transform;
-    const h = reportedH > 0 ? reportedH : transform ? textMatrixFontSize(transform) : 0;
-    const itemKey = textItemDedupeKey(item.str, w, h, transform, item.fontName);
-    const seenIndex = seenTextItems.get(itemKey);
-    if (seenIndex !== undefined) {
-      // Overprinted text often appears twice with identical geometry,
-      // sometimes differing only in pdf.js' hard-EOL flag. Keep one text
-      // run, but preserve the line-break signal if any duplicate carries it.
-      if (item.hasEOL) joinItems[seenIndex].hasEOL = true;
-      continue;
-    }
-    textArea += Math.abs(w * h);
-
-    // Feed the page-text joiner. x/fontSize default to 0 when the
-    // item lacks a transform (pdf.js does this for synthetic-EOL
-    // items); the joiner already handles zero fontSize by falling back
-    // to a neighbour.
-    const itemX = transform ? transform[4] : 0;
-    const itemFontSize = transform ? textMatrixFontSize(transform, h) : h;
-    seenTextItems.set(itemKey, joinItems.length);
-    joinItems.push({
-      str: item.str,
-      x: itemX,
-      width: w,
-      fontSize: itemFontSize,
-      hasEOL: !!item.hasEOL,
-      ...(typeof item.dir === 'string' && { dir: item.dir }),
-    });
-
-    // Skip whitespace-only items in spans output — pdf.js emits a span
-    // for every positioned space, which can double the array length and
-    // sometimes carries a synthetic width that exceeds the page width.
-    // The aggregate `text` already preserves the spaces, so layout
-    // analysis loses nothing; downstream agents get a cleaner signal.
-    if (wantSpans && item.str.trim().length > 0 && transform) {
-      const geometry = textRunGeometryFromTransform({
-        transform,
-        width: w,
-        height: h,
-        pageHeight: height,
-        viewMinX: xMin,
-        viewMinY: yMin,
-        dir: typeof item.dir === 'string' ? item.dir : undefined,
-      });
-      spans.push({
-        text: flags.normalize ? normalizeText(item.str) : item.str,
-        ...geometry,
-        ...(typeof item.fontName === 'string' && { fontName: item.fontName }),
-      });
-    }
-  }
-  const rawText = joinPageText(joinItems).trimEnd();
-  // charCount must reflect the string the caller actually receives, so
-  // measure after normalization.
-  const text = flags.normalize ? normalizeText(rawText) : rawText;
-  // Only surface rawText when normalization actually changed the string —
-  // exposing it unconditionally would double JSON size for the common
-  // case of already-canonical PDFs.
-  const preservedRaw = flags.normalize && rawText !== text ? rawText : undefined;
-
-  // Always expand image-bbox per instance — counting ops would under-
-  // report when pdf.js's QueueOptimizer collapses N draws of the same
-  // XObject into a single Repeat / Group op. Expanded boxes serve as
-  // both the public `imageBoxes` payload (when requested) and the source
-  // of `imageCount`, which keeps the two trivially consistent.
-  const opList = await page.getOperatorList();
-  const allBoxes = buildImageBoxes(opList.fnArray, opList.argsArray as unknown[][], ops, height, view[0], yMin);
-  const imageCount = allBoxes.length;
-  const imageBoxes = flags.imageBoxes ? allBoxes : undefined;
-  const vectorCount = countVectorPaintOps(
-    opList.fnArray,
-    opList.argsArray as unknown[][],
-    ops,
-    width,
-    height,
-    xMin,
-    yMin,
-  );
-  const allVectorBoxes =
-    vectorCount > 0
-      ? buildVectorBoxes(opList.fnArray, opList.argsArray as unknown[][], ops, width, height, xMin, yMin)
-      : [];
-  const vectorBoxes = flags.vectorBoxes ? allVectorBoxes : undefined;
-  // Build layout internally for form-field labels and visual-region table
-  // hints, but only expose pages[].layout when --layout is explicitly on.
-  const internalLayout =
-    flags.layout || flags.visualRegions || flags.formFields || flags.links || flags.needFormFieldsForSearch
-      ? buildLayout(spans, round2(width), round2(height))
-      : undefined;
-  const layout = flags.layout ? internalLayout : undefined;
-  const needsAnnotations =
-    flags.formFields ||
-    flags.links ||
-    flags.annotations ||
-    flags.visualRegions ||
-    flags.annotationAppearanceHints ||
-    flags.needAnnotationsForSearch ||
-    flags.needFormFieldsForSearch;
-  const annotations = needsAnnotations ? await page.getAnnotations({ intent: 'display' }) : undefined;
-  const visibleAnnotationAppearance = annotations ? hasVisibleAnnotationAppearance(annotations) : false;
-  const widgetAppearanceCaptions = annotations?.some(hasPushButtonWidget) ? getWidgetAppearanceCaptions?.() : undefined;
-  const allFormFields =
-    flags.formFields || flags.visualRegions || flags.needFormFieldsForSearch
-      ? buildFormFields(
-          annotations ?? [],
-          height,
-          xMin,
-          yMin,
-          flags.formFields || flags.visualRegions || flags.needFormFieldsForSearch
-            ? [
-                ...(internalLayout?.blocks.flatMap((block) =>
-                  (block.lines.length > 0 ? block.lines : [block]).map((item) => ({
-                    text: item.text,
-                    x: item.x,
-                    y: item.y,
-                    width: item.width,
-                    height: item.height,
-                    ...('fontSize' in item && item.fontSize !== undefined && { fontSize: item.fontSize }),
-                  })),
-                ) ?? []),
-                ...spans.map((span) => ({
-                  text: span.text,
-                  x: span.x,
-                  y: span.y,
-                  width: span.width,
-                  height: span.height,
-                  fontSize: span.fontSize,
-                })),
-              ]
-            : [],
-          { widgetAppearanceCaptions },
-        )
-      : undefined;
-  const formFields = flags.formFields ? allFormFields : undefined;
-  const internalFormFields = flags.needFormFieldsForSearch ? allFormFields : undefined;
-  const links = flags.links
-    ? await buildLinks(annotations ?? [], height, xMin, yMin, {
-        resolveDestinationPage: (target) => resolveDestinationPage(doc, target),
-        labelLines:
-          internalLayout?.blocks.flatMap((block) =>
-            (block.lines.length > 0 ? block.lines : [block]).map((item) => ({
-              text: item.text,
-              x: item.x,
-              y: item.y,
-              width: item.width,
-              height: item.height,
-            })),
-          ) ?? [],
-      })
-    : undefined;
-  const allPageAnnotations =
-    flags.annotations || flags.visualRegions || flags.needAnnotationsForSearch
-      ? buildAnnotations(annotations ?? [], height, xMin, yMin, {
-          normalizeText: flags.normalize ? normalizeText : undefined,
-        })
-      : undefined;
-  const pageAnnotations = flags.annotations ? allPageAnnotations : undefined;
-  const internalAnnotations = flags.needAnnotationsForSearch ? allPageAnnotations : undefined;
-  const structure = flags.structure
-    ? buildPageStructure(await page.getStructTree(), {
-        normalizeText: flags.normalize ? normalizeText : undefined,
-        pageHeight: height,
-        viewMinX: xMin,
-        viewMinY: yMin,
-      })
-    : undefined;
-  const jsActions = flags.viewer
-    ? normalizeJavaScriptActions(await page.getJSActions(), {
-        normalizeText: flags.normalize ? normalizeText : undefined,
-      })
-    : undefined;
-
-  const pageArea = width * height;
-  const rawCoverage = pageArea > 0 ? textArea / pageArea : 0;
-  const textCoverage = Math.max(0, Math.min(1, rawCoverage));
-  const rasterBackedTextLayer = isRasterBackedTextLayer({
-    imageCount,
-    vectorCount,
-    textCoverage,
-    charCount: text.length,
-    imageBoxes: allBoxes,
-    pageWidth: width,
-    pageHeight: height,
-  });
-
-  const visualRegionInput = flags.visualRegions
-    ? {
-        pageWidth: round2(width),
-        pageHeight: round2(height),
-        imageBoxes: allBoxes,
-        vectorBoxes: allVectorBoxes,
-        layout: internalLayout,
-        formFields: allFormFields,
-        annotations: allPageAnnotations,
-      }
-    : undefined;
-
-  // Measured on the text we actually return (post-normalize) so the
-  // count + ratio match what an agent sees in `text`. Cheap (one
-  // string walk), so always on — this is the primary signal for
-  // catching ToUnicode-CMap-less PDFs that look 100% covered but emit
-  // raw glyph indices. Surfacing the raw count alongside the ratio
-  // keeps sparse occurrences (a handful of control chars in a long
-  // body page) discriminable from "zero" when the 3dp ratio rounds
-  // them down.
-  const npStats = nonPrintableStats(text);
-
-  return {
-    text,
-    rawText: preservedRaw,
-    charCount: text.length,
-    imageCount,
-    rasterBackedTextLayer,
-    optionalContentText,
-    vectorCount,
-    textCoverage: Math.round(textCoverage * 1000) / 1000,
-    nonPrintableRatio: npStats.ratio,
-    nonPrintableCount: npStats.count,
-    // Round to 2dp; PDF dimensions are nominally integers (Letter 612×792,
-    // A4 595×842) but encrypted/cropped PDFs can carry sub-point fractions.
-    width: round2(width),
-    height: round2(height),
-    // Spans are only exposed publicly when --geometry is on; layout /
-    // imageBoxes each have their own opt-in flags and are independent
-    // of `geometry`. `_internalSpans` only rides along when search
-    // actually needs them — `--layout` alone already consumed the
-    // span list during `buildLayout` above, so re-emitting them on
-    // PageData would waste memory on the typical extraction.
-    ...(flags.geometry && { spans }),
-    ...(flags.needSpansForSearch && { _internalSpans: spans }),
-    ...(layout !== undefined && { layout }),
-    ...(imageBoxes !== undefined && { imageBoxes }),
-    _warningImageBoxes: allBoxes,
-    ...(vectorBoxes !== undefined && { vectorBoxes }),
-    _warningVectorBoxes: allVectorBoxes,
-    ...(visualRegionInput !== undefined && { _visualRegionInput: visualRegionInput }),
-    ...(visibleAnnotationAppearance && { hasVisibleAnnotationAppearance: true }),
-    ...(formFields !== undefined && { formFields }),
-    ...(internalFormFields !== undefined && { _internalFormFields: internalFormFields }),
-    ...(links !== undefined && { links }),
-    ...(pageAnnotations !== undefined && { annotations: pageAnnotations }),
-    ...(internalAnnotations !== undefined && { _internalAnnotations: internalAnnotations }),
-    ...(structure !== undefined && { structure }),
-    ...(jsActions !== undefined && { jsActions }),
-  };
-}
-
-function isOptionalContentTextMarker(item: unknown): boolean {
-  if (!item || typeof item !== 'object') return false;
-  const marker = item as { type?: unknown; tag?: unknown };
-  return marker.type === 'beginMarkedContentProps' && marker.tag === 'OC';
-}
 
 /**
  * Extract a structured representation of a PDF.
